Move ActivityItem back into redux-slices/ui
This type is more or less UI-only, so it doesn't belong in the top-level
`types.ts` file.
