+++ ui/components/Wallet/WalletActivityListItem.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from "react"
 import dayjs from "dayjs"
 import classNames from "classnames"
 import { convertToEth } from "@tallyho/tally-background/lib/utils"
-import { ActivityItem } from "@tallyho/tally-background/types"
+import { ActivityItem } from "@tallyho/tally-background/redux-slices/ui"
 
 interface Props {
   onClick: () => void
